requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
alembic==1.13.1
    # via -r requirements.in
annotated-types==0.7.0
    # via pydantic
anyio==4.10.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
argon2-cffi==25.1.0
    # via minio
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
async-timeout==5.0.1
    # via redis
bandit==1.7.6
    # via -r requirements.in
bcrypt==3.2.2
    # via
    #   -r requirements.in
    #   passlib
black==23.12.1
    # via -r requirements.in
build==1.3.0
    # via pip-tools
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   minio
cffi==2.0.0
    # via
    #   argon2-cffi-bindings
    #   bcrypt
    #   cryptography
click==8.1.8
    # via
    #   black
    #   pip-tools
    #   uvicorn
coverage[toml]==7.10.7
    # via
    #   coverage
    #   pytest-cov
cryptography==46.0.1
    # via python-jose
ecdsa==0.19.1
    # via python-jose
exceptiongroup==1.3.0
    # via
    #   anyio
    #   pytest
fastapi==0.109.0
    # via -r requirements.in
flake8==7.0.0
    # via -r requirements.in
gitdb==4.0.12
    # via gitpython
gitpython==3.1.45
    # via bandit
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.26.0
    # via -r requirements.in
idna==3.10
    # via
    #   anyio
    #   httpx
importlib-metadata==8.7.0
    # via build
iniconfig==2.1.0
    # via pytest
mako==1.3.10
    # via alembic
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via mako
mccabe==0.7.0
    # via flake8
mdurl==0.1.2
    # via markdown-it-py
minio==7.2.3
    # via -r requirements.in
mypy==1.8.0
    # via -r requirements.in
mypy-extensions==1.1.0
    # via
    #   black
    #   mypy
packaging==25.0
    # via
    #   black
    #   build
    #   pytest
passlib[bcrypt]==1.7.4
    # via
    #   -r requirements.in
    #   passlib
pathspec==0.12.1
    # via black
pillow==10.2.0
    # via -r requirements.in
pip-tools==7.3.0
    # via -r requirements.in
platformdirs==4.4.0
    # via black
pluggy==1.6.0
    # via pytest
psutil==6.1.1
    # via -r requirements.in
psycopg2-binary==2.9.9
    # via -r requirements.in
pyasn1==0.6.1
    # via
    #   python-jose
    #   rsa
pycodestyle==2.11.1
    # via flake8
pycparser==2.23
    # via cffi
pycryptodome==3.23.0
    # via minio
pydantic==2.5.3
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.14.6
    # via pydantic
pydantic-settings==2.1.0
    # via -r requirements.in
pyflakes==3.2.0
    # via flake8
pygments==2.19.2
    # via rich
pyproject-hooks==1.2.0
    # via build
pytest==7.4.4
    # via
    #   -r requirements.in
    #   pytest-asyncio
    #   pytest-cov
pytest-asyncio==0.23.3
    # via -r requirements.in
pytest-cov==4.1.0
    # via -r requirements.in
python-dotenv==1.0.0
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-jose[cryptography]==3.3.0
    # via
    #   -r requirements.in
    #   python-jose
python-multipart==0.0.6
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   bandit
    #   uvicorn
redis==5.0.1
    # via -r requirements.in
rich==14.1.0
    # via bandit
rsa==4.9.1
    # via python-jose
six==1.17.0
    # via ecdsa
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
sqlalchemy==2.0.25
    # via
    #   -r requirements.in
    #   alembic
starlette==0.35.1
    # via fastapi
stevedore==5.5.0
    # via bandit
tomli==2.2.1
    # via
    #   black
    #   build
    #   coverage
    #   mypy
    #   pip-tools
    #   pytest
typing-extensions==4.15.0
    # via
    #   alembic
    #   anyio
    #   black
    #   cryptography
    #   exceptiongroup
    #   fastapi
    #   gitpython
    #   minio
    #   mypy
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   starlette
    #   uvicorn
urllib3==2.5.0
    # via minio
uvicorn[standard]==0.27.0
    # via
    #   -r requirements.in
    #   uvicorn
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
wheel==0.45.1
    # via pip-tools
zipp==3.23.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
email-validator==2.1.0
pyodbc==5.0.1
apscheduler==3.10.4
